chore: adjust to protocols
