+++ substrafl/nodes/__init__.py
@@ -1,10 +1,13 @@
 from substrafl.nodes.schemas import OperationKey  # isort:skip
 
 from substrafl.nodes.aggregation_node import AggregationNode
+from substrafl.nodes.aggregation_node import SimuAggregationNode
 from substrafl.nodes.protocol import AggregationNodeProtocol
 from substrafl.nodes.protocol import TestDataNodeProtocol
 from substrafl.nodes.protocol import TrainDataNodeProtocol
+from substrafl.nodes.test_data_node import SimuTestDataNode
 from substrafl.nodes.test_data_node import TestDataNode
+from substrafl.nodes.train_data_node import SimuTrainDataNode
 from substrafl.nodes.train_data_node import TrainDataNode
 
 # This is needed for auto doc to find that Node module's is organizations.organization, otherwise when
@@ -18,5 +21,8 @@ __all__ = [
     "AggregationNode",
     "TrainDataNode",
     "TestDataNode",
+    "SimuAggregationNode",
+    "SimuTrainDataNode",
+    "SimuTestDataNode",
     "OperationKey",
 ]
