+++ substrafl/nodes/test_data_node.py
@@ -1,4 +1,7 @@
+from __future__ import annotations
+
 import uuid
+from typing import TYPE_CHECKING
 from typing import Dict
 from typing import List
 from typing import Optional
@@ -10,9 +13,14 @@ from substrafl.nodes.protocol import TestDataNodeProtocol
 from substrafl.nodes.schemas import InputIdentifiers
 from substrafl.nodes.schemas import OperationKey
 from substrafl.nodes.schemas import OutputIdentifiers
+from substrafl.nodes.schemas import SimuPerformancesMemory
+from substrafl.nodes.utils import preload_data
 from substrafl.remote.operations import RemoteDataOperation
 from substrafl.remote.register import register_function
 from substrafl.remote.remote_struct import RemoteStruct
+
+if TYPE_CHECKING:
+    from substrafl.compute_plan_builder import ComputePlanBuilder
 
 
 class TestDataNode(TestDataNodeProtocol):
@@ -148,6 +156,71 @@ class TestDataNode(TestDataNodeProtocol):
                 task["function_key"] = function_key
 
         return cache
+
+    def summary(self) -> dict:
+        """Summary of the class to be exposed in the experiment summary file
+
+        Returns:
+            dict: a json-serializable dict with the attributes the user wants to store
+        """
+        return {
+            "organization_id": self.organization_id,
+            "data_manager_key": self.data_manager_key,
+            "data_sample_keys": self.data_sample_keys,
+        }
+
+
+class SimuTestDataNode(TestDataNodeProtocol):
+    def __init__(self, client: substra.Client, node: TestDataNode, strategy: ComputePlanBuilder):
+        self.organization_id = node.organization_id
+
+        self.data_manager_key = node.data_manager_key
+        self.data_sample_keys = node.data_sample_keys
+
+        self._datasamples = preload_data(
+            client=client, data_manager_key=self.data_manager_key, data_sample_keys=self.data_sample_keys
+        )
+
+        self._strategy = strategy
+        self._memory = SimuPerformancesMemory()
+
+    def update_states(
+        self,
+        operation: RemoteDataOperation,
+        round_idx: Optional[int] = None,
+        *args,
+        **kwargs,
+    ) -> None:
+        """This function will execute the method to run on the test node with the argument
+        `_skip=True`, to execute it directly in RAM. The executed method is expected to compute
+        metrics, and to return a dictionary with the metric name as key and the result as value.
+        It will also stores the computed scores in a `SimulationPerformance` object.
+
+        This function is expected to overload the `update_state` method of a TestDataNode
+        to simulate its execution on RAM only.
+
+        Args:
+            operation (RemoteDataOperation): Object containing all the information to execute
+                the method.
+            round_idx (Optional[int]): Current round idx. Defaults to None.
+        """
+        method_name = operation.remote_struct._method_name
+        method_parameters = operation.remote_struct._method_parameters
+
+        method_parameters["shared_state"] = operation.shared_state
+        method_parameters["datasamples"] = self._datasamples
+        method_to_run = getattr(self._strategy, method_name)
+
+        scores = method_to_run(**method_parameters, _skip=True)
+
+        for metric in scores:
+            self._memory.performance.append(scores[metric])
+            self._memory.identifier.append(metric)
+            self._memory.round_idx.append(round_idx)
+            self._memory.worker.append(self.organization_id)
+
+    def register_operations(self, *args, **kwargs) -> None:
+        return
 
     def summary(self) -> dict:
         """Summary of the class to be exposed in the experiment summary file
