--- substrafl/simulation/__init__.py
@@ -1,13 +1,0 @@
-from substrafl.simulation.nodes import SimuAggregationNode
-from substrafl.simulation.nodes import SimuTestDataNode
-from substrafl.simulation.nodes import SimuTrainDataNode
-from substrafl.simulation.schemas import SimuPerformancesMemory
-from substrafl.simulation.schemas import SimuStatesMemory
-
-__all__ = [
-    "SimuTrainDataNode",
-    "SimuTestDataNode",
-    "SimuAggregationNode",
-    "SimuPerformancesMemory",
-    "SimuStatesMemory",
-]
